+++ spec/generator_spec.rb
@@ -204,7 +204,7 @@ describe CloudCannonJekyll::Generator do
         expect(config["paths"]["collections"]).to be_nil
         expect(config["paths"]["includes"]).to be_nil
         expect(config["paths"]["layouts"]).to be_nil
-      elsif Jekyll::VERSION.match? %r!3\.[0-4]\.!
+      elsif Jekyll::VERSION.match? %r!3\.[0-5]\.!
         expect(config["paths"]["plugins"]).to eq("_plugins")
         expect(config["paths"]["data"]).to eq("_data")
         expect(config["paths"]["collections"]).to be_nil
@@ -252,6 +252,26 @@ describe CloudCannonJekyll::Generator do
         "things" => ["hello", "there"],
         "staff" => ["jim", "bob"]
       })
+    end
+
+    it "has populated input-options" do
+      expect(config["input-options"]).to eq({
+        "content" => {"image" => true, "bold" => true},
+        "my_html" => {"italic" => true, "bold" => true, "styles" => "hello.css"}
+      })
+    end
+
+    it "has populated defaults" do
+      expect(config["defaults"]).to eq([
+        {
+          "scope" => {"path" => ""},
+          "values" => {"layout" => "page"}
+        },
+        {
+          "scope" => {"path" => "", "type" => "posts"},
+          "values" => {"layout" => "post"}
+        }
+      ])
     end
   end
 
@@ -317,6 +337,14 @@ describe CloudCannonJekyll::Generator do
 
     it "has no select-data" do
       expect(config).not_to have_key("select-data")
+    end
+
+    it "has no input-options" do
+      expect(config).not_to have_key("input-options")
+    end
+
+    it "has no defaults" do
+      expect(config["defaults"]).to eq([])
     end
   end